[fix] Resolve GCC build warnings (#538)
* resolve gcc warnings

* remove unused macro

* correct some type declaration

* add missing space

* Make sure fallthrough attribute is specific to GCC

* remove useless code and avoid strict aliasing warning

* move MAYBE_UNUSED macro to header file

* Revert "remove useless code and avoid strict aliasing warning"

This reverts commit 403ee5e43770894f2b5303b0c11e9dd0873f892a.

+++ sources/isal/crc/crc_base.c
@@ -244,7 +244,7 @@ static const uint32_t crc32_table_gzip_refl[256] = {
 
 uint16_t crc16_t10dif_base(uint16_t seed, uint8_t * buf, uint64_t len)
 {
-	int i;
+	uint64_t i;
 	uint16_t crc = seed;
 
 	for (i = 0; i < len; i++)
@@ -255,7 +255,7 @@ uint16_t crc16_t10dif_base(uint16_t seed, uint8_t * buf, uint64_t len)
 
 uint16_t crc16_t10dif_copy_base(uint16_t seed, uint8_t * dst, uint8_t * src, uint64_t len)
 {
-	int i;
+	uint64_t i;
 	uint16_t crc = seed;
 
 	for (i = 0; i < len; i++) {
